Module google_tasks1::api

Structs

Task
There is no detailed description.

TaskClearCall
Clears all completed tasks from the specified task list. The affected tasks will be marked as 'hidden' and no longer be returned by default when retrieving all tasks for a task list.

TaskDeleteCall
Deletes the specified task from the task list.

TaskGetCall
Returns the specified task.

TaskInsertCall
Creates a new task on the specified task list.

TaskLinks
Collection of links. This collection is read-only.

TaskList
There is no detailed description.

TaskListCall
Returns all tasks in the specified task list.

TaskLists
There is no detailed description.

TaskMethods
A builder providing access to all methods supported on task resources. It is not used directly, but through the TasksHub hub.

TaskMoveCall
Moves the specified task to another position in the task list. This can include putting it as a child task under a new parent and/or move it to a different position among its sibling tasks.

TaskPatchCall
Updates the specified task. This method supports patch semantics.

TaskUpdateCall
Updates the specified task.

TasklistDeleteCall
Deletes the authenticated user's specified task list.

TasklistGetCall
Returns the authenticated user's specified task list.

TasklistInsertCall
Creates a new task list and adds it to the authenticated user's task lists.

TasklistListCall
Returns all the authenticated user's task lists.

TasklistMethods
A builder providing access to all methods supported on tasklist resources. It is not used directly, but through the TasksHub hub.

TasklistPatchCall
Updates the authenticated user's specified task list. This method supports patch semantics.

TasklistUpdateCall
Updates the authenticated user's specified task list.

Tasks
There is no detailed description.

TasksHub
Central instance to access all TasksHub related resource activities

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.