Struct google_tasks1::TaskMethods

pub struct TaskMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on task resources. It is not used directly, but through the TasksHub hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_tasks1 as tasks1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use tasks1::TasksHub;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = TasksHub::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `clear(...)`, `delete(...)`, `get(...)`, `insert(...)`, `list(...)`, `move_(...)`, `patch(...)` and `update(...)`
// to build up your call.
let rb = hub.tasks();

Methods

impl<'a, C, A> TaskMethods<'a, C, A>

fn list(&self, tasklist: &str) -> TaskListCall<'a, C, A>

Create a builder to help you perform the following task:

Returns all tasks in the specified task list.

Arguments

tasklist - Task list identifier.

fn update(&self, request: Task, tasklist: &str, task: &str) -> TaskUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Updates the specified task.

Arguments

request - No description provided.

tasklist - Task list identifier.

task - Task identifier.

fn patch(&self, request: Task, tasklist: &str, task: &str) -> TaskPatchCall<'a, C, A>

Create a builder to help you perform the following task:

Updates the specified task. This method supports patch semantics.

Arguments

request - No description provided.

tasklist - Task list identifier.

task - Task identifier.

fn clear(&self, tasklist: &str) -> TaskClearCall<'a, C, A>

Create a builder to help you perform the following task:

Clears all completed tasks from the specified task list. The affected tasks will be marked as 'hidden' and no longer be returned by default when retrieving all tasks for a task list.

Arguments

tasklist - Task list identifier.

fn move_(&self, tasklist: &str, task: &str) -> TaskMoveCall<'a, C, A>

Create a builder to help you perform the following task:

Moves the specified task to another position in the task list. This can include putting it as a child task under a new parent and/or move it to a different position among its sibling tasks.

Arguments

tasklist - Task list identifier.

task - Task identifier.

fn delete(&self, tasklist: &str, task: &str) -> TaskDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes the specified task from the task list.

Arguments

tasklist - Task list identifier.

task - Task identifier.

fn get(&self, tasklist: &str, task: &str) -> TaskGetCall<'a, C, A>

Create a builder to help you perform the following task:

Returns the specified task.

Arguments

tasklist - Task list identifier.

task - Task identifier.

fn insert(&self, request: Task, tasklist: &str) -> TaskInsertCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a new task on the specified task list.

Arguments

request - No description provided.

tasklist - Task list identifier.