Struct google_tasks1::TasklistMethods

pub struct TasklistMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on tasklist resources. It is not used directly, but through the TasksHub hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_tasks1 as tasks1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use tasks1::TasksHub;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = TasksHub::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `delete(...)`, `get(...)`, `insert(...)`, `list(...)`, `patch(...)` and `update(...)`
// to build up your call.
let rb = hub.tasklists();

Methods

impl<'a, C, A> TasklistMethods<'a, C, A>

fn list(&self) -> TasklistListCall<'a, C, A>

Create a builder to help you perform the following task:

Returns all the authenticated user's task lists.

fn update(&self, request: TaskList, tasklist: &str) -> TasklistUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Updates the authenticated user's specified task list.

Arguments

request - No description provided.

tasklist - Task list identifier.

fn delete(&self, tasklist: &str) -> TasklistDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes the authenticated user's specified task list.

Arguments

tasklist - Task list identifier.

fn patch(&self, request: TaskList, tasklist: &str) -> TasklistPatchCall<'a, C, A>

Create a builder to help you perform the following task:

Updates the authenticated user's specified task list. This method supports patch semantics.

Arguments

request - No description provided.

tasklist - Task list identifier.

fn insert(&self, request: TaskList) -> TasklistInsertCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a new task list and adds it to the authenticated user's task lists.

Arguments

request - No description provided.

fn get(&self, tasklist: &str) -> TasklistGetCall<'a, C, A>

Create a builder to help you perform the following task:

Returns the authenticated user's specified task list.

Arguments

tasklist - Task list identifier.