Struct google_tasks1::Task

pub struct Task {
    pub status: Option<String>,
    pub kind: Option<String>,
    pub parent: Option<String>,
    pub links: Option<Vec<TaskLinks>>,
    pub title: Option<String>,
    pub deleted: Option<bool>,
    pub completed: Option<String>,
    pub updated: Option<String>,
    pub due: Option<String>,
    pub etag: Option<String>,
    pub id: Option<String>,
    pub position: Option<String>,
    pub hidden: Option<bool>,
    pub notes: Option<String>,
    pub self_link: Option<String>,
}

There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list tasks (none)

update tasks (request|response)

patch tasks (request|response)

clear tasks (none)

move tasks (response)

delete tasks (none)

get tasks (response)

insert tasks (request|response)

Fields
status: Option<String>

Status of the task. This is either "needsAction" or "completed".

kind: Option<String>

Type of the resource. This is always "tasks#task".

parent: Option<String>

Parent task identifier. This field is omitted if it is a top-level task. This field is read-only. Use the "move" method to move the task under a different parent or to the top level.

links: Option<Vec<TaskLinks>>

Collection of links. This collection is read-only.

title: Option<String>

Title of the task.

deleted: Option<bool>

Flag indicating whether the task has been deleted. The default if False.

completed: Option<String>

Completion date of the task (as a RFC 3339 timestamp). This field is omitted if the task has not been completed.

updated: Option<String>

Last modification time of the task (as a RFC 3339 timestamp).

due: Option<String>

Due date of the task (as a RFC 3339 timestamp). Optional.

etag: Option<String>

ETag of the resource.

id: Option<String>

Task identifier.

position: Option<String>

String indicating the position of the task among its sibling tasks under the same parent task or at the top level. If this string is greater than another task's corresponding position string according to lexicographical ordering, the task is positioned after the other task under the same parent task (or at the top level). This field is read-only. Use the "move" method to move the task to another position.

hidden: Option<bool>

Flag indicating whether the task is hidden. This is the case if the task had been marked completed when the task list was last cleared. The default is False. This field is read-only.

notes: Option<String>

Notes describing the task. Optional.

self_link: Option<String>

URL pointing to this task. Used to retrieve, update, or delete this task.